requirements.txt
streamlit
torch
torchvision
pillow
pytesseract
transformers
numpy
opencv-python
gdown
